fix warning

+++ mseg/utils/conn_comp.py
@@ -30,7 +30,7 @@ def scipy_conn_comp(img: np.ndarray) -> Dict[int,List[np.ndarray]]:
 	present_class_idxs = np.unique(img)
 
 	for class_idx in present_class_idxs:
-		structure = np.ones((3, 3), dtype=np.int)  # this defines the connection filter
+		structure = np.ones((3, 3), dtype=np.int32)  # this defines the connection filter
 		instance_img, nr_objects = scipy.ndimage.label(input=img == class_idx, structure=structure)
 		
 		for i in np.unique(instance_img):
